New Version
Decided not to end the project here. So far I've added an option to
unmirror the image. The webcam also displays the image taken and takes
the picture in 1080p.

I still have to adjust the size of the snapshot window. I'm also adding
an option to change the size of the picture and I'm going to adjust the
layout of the panel with sizers instead of statically choosing it.